+++ docker-compose.yml
@@ -10,7 +10,7 @@ services:
     labels:
       lagoon.type: node
     volumes:
-      - .:/app:delegated
+      - ./src:/app/src:delegated
     environment:
       - LAGOON_LOCALDEV_HTTP_PORT=3000
       - LAGOON_ROUTE=http://node.docker.amazee.io
