--- index.js
@@ -1,16 +1,0 @@
-const express = require('express')
-const app = express()
-
-app.get('/', function (req, res) {
-  let result = []
-  Object.keys(process.env).map(key => {
-    result.push(`${key}=${process.env[key]}`)
-  })
-  result.sort()
-  res.header('X-LAGOON' , process.env.HOSTNAME )
-  res.send(result.join("<br />"))
-})
-
-app.listen(3000, function () {
-  console.log('Example app listening on port 3000!')
-})
